Fix debug checks when unstuffing packets
In packet::framing::unstuff(), the code may be larger than the remaining
number of bytes left in the packet. This is because the code denotes the
location of the next zero and may mark the end-of-packet zero byte. The
the length parameter to the unstuff() function is equal to the number of
bytes to unstuff, excluding the end-of-packet zero byte.