Set single-byte charset in gdbinit
The default character set used by gdb is often set to UTF-8. As we
frequently use uint8_t to store values that are NOT valid UTF-8 (a
variable-length encoding), this will cause gdb to crash when printing
invalid character constants.

In gdbinit, use ISO-8859-1 single-byte character set as host and target
charset.