Add periodic encoder events
In the situation where a large amount of time without an encoder event,
the position estimate from the best fit polynomial will have a large
error. In this case, we return the last observed count instead of the
estimated value. To mitigate this, we start a deadline timer to add a
new entry to the encoder event buffer if no edges are detected for 10
ms.