Fix increment of event index
The event index points to the oldest entry in the event buffer. When
using 'skip', the newest entry may need to be overwritten as we 'skip' a
number of encoder events before saving them to the buffer and advancing
to next index. However, we always include the most recent event for
estimation purposes.

If we have not finished incrementing the skip counter for a full cycle,
decrement the event index before writing to the event buffer in order to
overwrite the newest event.