Fix rear wheel angle and forward velocity calculations
Include the roller to rear wheel ratio in RLS_GTS35_ENC_CFG to give the
encoder count for the rear wheel directly. Increase counts per
revolution by 4x as 48 is the number of slots and not equal to counts
when the encoder is used in quadrature.

Forward velocity v is positive for negative rear wheel angular velocity
and is redefined accordingly.