Handle position overflow in FOAH velocity estimation
Add new argument 'overflow_value' to foah::estimate_velocity() with
default value of 0. This argument represents a positive value with would
cause the underlying data to overflow; this commonly occurs with
encoders. If positive, the position buffer elements will be modified in
place if overflow or underflow is detected during the velocity
estimation calculation. The buffer elements will be restored at the end.