Decimate mpl pose plots
Matplotlib plots of the time series pose fields are very slow due to the
large number of samples drawn on screen. To improve performance,
decimate plots based on the x-axis limits to keep constant the samples
drawn. Decimation factor is displayed in the plot title.

This commit disables plotting of the rear wheel trajectory.