Limit pitch constraint iterations
Limit Newton-Raphson iterations when solving for pitch angle constraint
to 3. In normal cases, the pitch angle can be solved in 1 to 2
constraints. It is only in extreme/unrealistic configurations where the
number of iterations increases, but these do not result in realistic
pitch values, so there is no reason to expend computation time.