Define a simulation protobuf message
Define a simulation protobuf messages instead of an application specific
message. This message encapsulates the data associated with a
sim::Bicycle object.

The functions in messageutil have been redefined to set a simulation
message instead of a clustril message.

Add an accessor to obtain the full state for a sim::Bicycle object.