Use COBS packet framing with larger payloads
Rename COBS_MAX_SIZE_DATA_SET to COBS_MAX_SIZE_PAYLOAD. Allow use of
COBS packet framing with payloads greater than 254 bytes. Larger
payloads are encoded a section of a time, using a section size of 254
bytes except for the last section.

Decoding allows for a buffer with more than 254 elements. The packet
delimiters must be present in the data to be decoded in this case.