Convert protobuf messages to numpy records
Create numpy record datatype from a protobuf message. If repeated fields
are used or a string/bytes field type is used, a nanopb options file
must de defined with the max count and max size set.

In numpy records, string and bytes type fields are redefined to be
arrays of np.uint8.