:wrench: Change meaning of COBS result object fields
EncodeResult.written becomes EncodeResult.produced. Its value is only useful
when the status is OK so in other cases it is set to 0.

DecodeResult.read becomes DecodeResult.consumed. Its value is only useful
when the status is OK or UNEXPECTED_ZERO so in other cases it is set to 0. This
makes things slightly easier for code calling `cobs::decode`.

DecodeResult.written becomes DecodeResult.produced. Its value is only useful
when the status is OK so in other cases it is set to 0.