Store computed input in sim::Bicycle
The input vector u is calculated from torque sensor measurements and
inertia properties of the system. This adds a member variable to
sim::Bicycle to store and access the most recently computed input,
allowing code calling the update_dynamics() member function to obtain
the roll torque and steer torque used in model simulation.