Add calculated feedforward torque to plots
Only commanded torque is recorded which is the sum of calculated
feedback and feedforward torques. As the value is quantized, the exact
values for feedback and feedforward torques cannot be determined and
instead are estimated. An additional feedforward term is calculated
using the state space equations and recorded state and input vectors.