Define USB config for project messaging
Define USBConfig struct in txrx_common for use with sending and
receiving protobuf messages. This is done to allow the data receive
callback to interact with the Receiver class. Starting the USB driver is
now handled by the Transmitter and Receiver class and is order
independent as long as both Receiver and Transmitter are instantiated
before start is called.